no-byte

+++ backend/src/PokeGame.Infrastructure/Converters/FriendshipConverter.cs
@@ -6,7 +6,7 @@ internal class FriendshipConverter : JsonConverter<Friendship>
 {
   public override Friendship? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetByte(out byte value) ? new Friendship(value) : null;
+    return reader.TryGetInt32(out int value) ? new Friendship(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, Friendship friendship, JsonSerializerOptions options)
